Revert "Dev (#16)" (#17)
This reverts commit 4245a67ece792143a1973c6e680ef5411cb1f258.

+++ src/main/java/com/hims/entity/BillingPolicyMaster.java
@@ -3,7 +3,6 @@ package com.hims.entity;
 import jakarta.persistence.*;
 import lombok.*;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Entity
@@ -33,7 +32,7 @@ public class BillingPolicyMaster {
     private Integer followupDaysAllowed;
 
     @Column(name = "discount_percentage")
-    private BigDecimal discountPercentage;
+    private Integer discountPercentage;
 
 //    @Column(name = "status", length = 1)
 //    private String status;
