+++ src/main/java/com/hims/request/BillingPolicyRequest.java
@@ -3,8 +3,6 @@ package com.hims.request;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.math.BigDecimal;
-
 @Getter
 @Setter
 public class BillingPolicyRequest {
@@ -12,5 +10,5 @@ public class BillingPolicyRequest {
     private String description;
     private String applicableBillingType;
     private Integer followupDaysAllowed;
-    private BigDecimal discountPercentage;
+    private Integer discountPercentage;
 }
